BloodRegistry/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==1.12.3
asgiref==3.4.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.3.0
base58==2.1.1
bitarray==1.2.2
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Cython==0.29.26
cytoolz==0.11.2
Django==3.2.3
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
haversine==2.6.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pathspec==0.9.0
pipx==0.17.0
platformdirs==2.4.1
Pillow==9.5
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
py-solc-x==1.1.1
pyasn1==0.4.8
pycryptodome==3.12.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dotenv==0.19.2
pytz==2022.2.1
requests==2.26.0
rlp==2.0.1
rsa==4.8
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
typed-ast==1.5.4
typing-extensions==4.0.1
urllib3==1.26.7
userpath==1.7.0
varint==1.0.2
web3==5.25.0
websockets==9.1
yarl==1.7.2
zipp==3.8.1